Update (user maintenance) user previews and basic pop ups
